SymbolicNeuralNetworks

Documentation for SymbolicNeuralNetworks.

SymbolicNeuralNetworks.symbolic_hamiltonian

SymbolicNeuralNetworks.symbolic_lagrangian

SymbolicNeuralNetworks.symbolic_hamiltonian — Method

symbolic_hamitonian is a functions that creates a symbolic hamiltonian from any hamiltonian.
The output of the function is 
    - the symbolics parameters used to build the hamiltonian (i.e t, q, p),
    - the symbolic expression of the hamiltonian,
    - the function generated from the symbolic hamiltonian.

SymbolicNeuralNetworks.symbolic_lagrangian — Method

symbolic_lagrangian is a functions that creates a symbolic lagrangian from any lagrangian.
The output of the function is 
    - the symbolics parameters used to build the hamiltonian (i.e t, x, v),
    - the symbolic expression of the lagrangian,
    - the function generated from the symbolic lagrangian.