SymbolicNeuralNetworks

SymbolicNeuralNetworks is a library for creating symbolic representations of relatively small neural networks on whose basis more complicated expressions can be build. It should mostly be used together with other packages like GeometricMachineLearning and GeometricIntegrators.

SymbolicNeuralNetworks.Derivative

SymbolicNeuralNetworks.FlatSlice

SymbolicNeuralNetworks.Gradient

SymbolicNeuralNetworks.Jacobian

SymbolicNeuralNetworks.SymbolicNeuralNetwork

SymbolicNeuralNetworks.SymbolicPullback

SymbolicNeuralNetworks._build_nn_function

SymbolicNeuralNetworks._build_nn_function

SymbolicNeuralNetworks._build_nn_function_iip

SymbolicNeuralNetworks._build_nn_function_iip

SymbolicNeuralNetworks._build_nn_function_per_leaf

SymbolicNeuralNetworks._float_if_integer

SymbolicNeuralNetworks._get_contents

SymbolicNeuralNetworks._get_params

SymbolicNeuralNetworks._iip_batch_wrapper

SymbolicNeuralNetworks._iip_batch_wrapper2

SymbolicNeuralNetworks._modify_integer

SymbolicNeuralNetworks._modify_integer2

SymbolicNeuralNetworks._oop_batch_wrapper

SymbolicNeuralNetworks._oop_batch_wrapper2

SymbolicNeuralNetworks._reduce

SymbolicNeuralNetworks._reduce_iip

SymbolicNeuralNetworks.allocate_batch_output

SymbolicNeuralNetworks.allocate_single_output

SymbolicNeuralNetworks.apply_element_wise

SymbolicNeuralNetworks.apply_element_wise

SymbolicNeuralNetworks.build_function_generated

SymbolicNeuralNetworks.build_nn_function

SymbolicNeuralNetworks.build_nn_function

SymbolicNeuralNetworks.build_nn_function

SymbolicNeuralNetworks.build_nn_function

SymbolicNeuralNetworks.derivative

SymbolicNeuralNetworks.fix_create_array

SymbolicNeuralNetworks.fix_map_reduce

SymbolicNeuralNetworks.flatten_eqs

SymbolicNeuralNetworks.function_valued_parameters

SymbolicNeuralNetworks.make_kernel

SymbolicNeuralNetworks.make_kernel2

SymbolicNeuralNetworks.make_kernel_iip

SymbolicNeuralNetworks.make_kernel_iip2

SymbolicNeuralNetworks.modify_input_arguments

SymbolicNeuralNetworks.modify_input_arguments2

SymbolicNeuralNetworks.modify_input_arguments_iip

SymbolicNeuralNetworks.modify_input_arguments_iip2

SymbolicNeuralNetworks.promoted_eltype

SymbolicNeuralNetworks.redirect_output_writes

SymbolicNeuralNetworks.rewrite_arguments

SymbolicNeuralNetworks.rewrite_arguments2

SymbolicNeuralNetworks.symbolic_pullback

SymbolicNeuralNetworks.symboliccounter!

SymbolicNeuralNetworks.symbolize!

SymbolicNeuralNetworks.unflatten

SymbolicNeuralNetworks.Derivative — Type

Derivative

SymbolicNeuralNetworks.FlatSlice — Type

FlatSlice(range, size)

Where an entry of an equation set ended up in the flat vector that flatten_eqs produces, and which shape it has to be given again by unflatten.

SymbolicNeuralNetworks.Gradient — Type

Gradient <: Derivative

Computes and stores the gradient of a symbolic function with respect to the parameters of a SymbolicNeuralNetwork.

Constructors

Gradient(f, nn)

Differentiate the symbolic f with respect to the parameters of nn.

Gradient(nn)

Compute the symbolic output of nn and differentiate it with respect to the parameters of nn. This does:

nn.model(nn.input, params(nn))

Examples

using SymbolicNeuralNetworks: SymbolicNeuralNetwork, Gradient, derivative
using AbstractNeuralNetworks

c = Chain(Dense(2, 1, tanh))
nn = SymbolicNeuralNetwork(c)
(Gradient(nn) |> derivative)[1].L1.b

# output

1-element Vector{Symbolics.Num}:
 1 - (tanh(W_2[1] + W_1[1, 1]*sinput[1] + W_1[1, 2]*sinput[2])^2)

Implementation

Internally the constructors are using symbolic_pullback.

SymbolicNeuralNetworks.Jacobian — Type

Jacobian <: Derivative

An subtype of Derivative. Computes the derivatives of a neural network with respect to its inputs.

Constructors

Jacobian(f, nn)
Jacobian(nn)

Compute the jacobian of a SymbolicNeuralNetwork with respect to the input arguments.

Keys

Jacobian has the following keys:

nn::SymbolicNeuralNetwork,

f: a symbolic expression to be differentiated,

□: a symbolic expression of the Jacobian.

If f is not supplied as an input argument than it is taken to be:

f = nn.model(nn.input, params(nn))

Implementation

For a function $f:\mathbb{R}^n\to\mathbb{R}^m$ we choose the following convention for the Jacobian:

\[\square_{ij} = \frac{\partial}{\partial{}x_j}f_i, \text{ i.e. } \square \in \mathbb{R}^{m\times{}n}\]

This is also used by Zygote and ForwardDiff.

Examples

Here we compute the Jacobian of a single-layer neural network $x \to \mathrm{tanh}(Wx + b)$. Its element-wise derivative is:

\[ \frac{\partial}{\partial_i}\sigma(\sum_{k}w_{jk}x_k + b_j) = \sigma'(\sum_{k}w_{jk}x_k + b_j)w_{ji}.\]

Also note that for this calculation $\mathrm{tanh}(x) = \frac{e^{2x} - 1}{e^{2x} + 1}$ and $\mathrm{tanh}'(x) = \frac{4e^{2x}}{(e^{2x} + 1)^2}.$

We can use Jacobian together with build_nn_function:

using SymbolicNeuralNetworks
using SymbolicNeuralNetworks: Jacobian, derivative
using AbstractNeuralNetworks: Dense, Chain, NeuralNetwork, params
using Symbolics
import Random

Random.seed!(123)

input_dim = 5
output_dim = 2
d = Dense(input_dim, 2, tanh)
c = Chain(d)
nn = SymbolicNeuralNetwork(c)
□ = SymbolicNeuralNetworks.Jacobian(nn)
# here we need to access the derivative and convert it into a function
jacobian1 = build_nn_function(derivative(□), nn)
ps = params(NeuralNetwork(c, Float64))
input = rand(input_dim)
#derivative
Dtanh(x::Real) = 4 * exp(2 * x) / (1 + exp(2x)) ^ 2
analytic_jacobian(i, j) = Dtanh(sum(k -> ps.L1.W[j, k] * input[k], 1:input_dim) + ps.L1.b[j]) * ps.L1.W[j, i]
jacobian1(input, ps) ≈ [analytic_jacobian(i, j) for j ∈ 1:output_dim, i ∈ 1:input_dim]

# output

true

SymbolicNeuralNetworks.SymbolicNeuralNetwork — Type

SymbolicNeuralNetwork <: AbstractSymbolicNeuralNetwork

A symbolic neural network realizes a symbolic represenation (of small neural networks).

Fields

The struct has the following fields:

architecture: the neural network architecture,

model: the model (typically a Chain that is the realization of the architecture),

params: the symbolic parameters of the network.

sinput: the symbolic input of the network.

Constructors

SymbolicNeuralNetwork(nn)

Make a SymbolicNeuralNetwork based on a AbstractNeuralNetworks.Network.

SymbolicNeuralNetworks.SymbolicPullback — Type

SymbolicPullback <: AbstractPullback

SymbolicPullback computes the symbolic pullback of a loss function.

Examples

using SymbolicNeuralNetworks
using AbstractNeuralNetworks
using AbstractNeuralNetworks: params
import Random
Random.seed!(123)

c = Chain(Dense(2, 1, tanh))
nn = NeuralNetwork(c)
snn = SymbolicNeuralNetwork(nn)
loss = FeedForwardLoss()
pb = SymbolicPullback(snn, loss)
ps = params(nn)
typeof(pb(ps, nn.model, (rand(2), rand(1)))[2](1))

# output

@NamedTuple{L1::@NamedTuple{W::Matrix{Float64}, b::Vector{Float64}}}

Keyword Arguments

cse: perform common subexpression elimination when generating code (default true). This matters a lot here: without it every one of the 2 * n_layers generated blocks re-emits the entire forward pass, which makes the code for networks with more than one hidden layer intractably large. See _build_nn_function.

inplace: evaluate a batch with an in-place kernel (default true). The pullback is the end of the differentiation chain, so nothing differentiates through it and the default is what you want here; inplace = false exists for symmetry with build_nn_function.

Implementation

An instance of SymbolicPullback stores

loss: an instance of a NetworkLoss,

fun: a function that is used to compute the pullback.

If we call the functor of an instance of SymbolicPullback on model, ps and input it returns:

_pullback.loss(model, ps, input...), _pullback.fun(input..., ps)

where the second output argument is again a function.

Extended help

We note the following seeming peculiarity:

using SymbolicNeuralNetworks
using AbstractNeuralNetworks: Chain, Dense, NeuralNetwork, FeedForwardLoss, params
using Symbolics
import Random
Random.seed!(123)

c = Chain(Dense(2, 1, tanh))
nn = NeuralNetwork(c)
snn = SymbolicNeuralNetwork(nn)
loss = FeedForwardLoss()
pb = SymbolicPullback(snn, loss)
input_output = (rand(2), rand(1))
loss_and_pullback = pb(params(nn), nn.model, input_output)
# note that we apply the second argument to another input `1`
pb_values = loss_and_pullback[2](1)

@variables soutput[1:SymbolicNeuralNetworks.output_dimension(nn.model)]
symbolic_pullbacks = SymbolicNeuralNetworks.symbolic_pullback(loss(nn.model, params(snn), snn.input, soutput), snn)
pb_values2 = build_nn_function(symbolic_pullbacks, params(snn), snn.input, soutput; reduce = +)(input_output[1], input_output[2], params(nn))

pb_values == (pb_values2 |> SymbolicNeuralNetworks._get_contents |> SymbolicNeuralNetworks._get_params)

# output

true

See the docstrings for symbolic_pullback, build_nn_function, _get_params and _get_contents for more info on the functions that we used here. The noteworthy thing in the expression above is that the functor of SymbolicPullback returns two objects: the first one is the loss value evaluated for the relevant parameters and inputs. The second one is a function that takes again an input argument and then finally returns the partial derivatives. But why do we need this extra step with another function?

Reverse Accumulation

In machine learning we typically do reverse accumulation to perform automatic differentiation (AD). Assuming we are given a function that is the composition of simpler functions $f = f_1\circ{}f_2\circ\cdots\circ{}f_n:\mathbb{R}^n\to\mathbb{R}^m$ reverse differentiation starts with output sensitivities and then successively feeds them through $f_n$, $f_{n-1}$ etc. So it does:

\[(\nabla_xf)^T = (\nabla_{x}f_1)^T(\nabla_{f_1(x)}f_2)^T\cdots(\nabla_{f_{n-1}(\cdots{}x)}f_n)^T(do),\]

where $do\in\mathbb{R}^m$ are the output sensitivities and the jacobians are stepwise multiplied from the left. So we propagate from the output stepwise back to the input. If we have $m=1$, i.e. if the output is one-dimensional, then the output sensitivities may simply be taken to be $do = 1$.

So in theory we could leave out this extra step: returning an object (that is stored in pb.fun) can be seen as unnecessary as we could simply store the equivalent of pb.fun(1.) in an instance of SymbolicPullback. It is however customary for a pullback to return a callable function (that depends on the output sensitivities), which is why we also choose to do this here, even if the output sensitivities are a scalar quantity.

SymbolicNeuralNetworks._build_nn_function — Method

_build_nn_function(eq, params, sinput, soutput)

Build a function that can process a matrix. See build_nn_function(::EqT, ::NeuralNetworkParameters, ::Symbolics.Arr).

Keyword Arguments

cse: perform common subexpression elimination (default true). See _build_nn_function(::EqT, ::NeuralNetworkParameters, ::Symbolics.Arr).

Implementation

Note that we have two input arguments here which means this method processes code differently than _build_nn_function(::EqT, ::NeuralNetworkParameters, ::Symbolics.Arr, ::Symbolics.Arr). Here we call:

fix_create_array,

rewrite_arguments2,

modify_input_arguments2,

fix_map_reduce.

See the docstrings for those functions for details on how the code is modified.

SymbolicNeuralNetworks._build_nn_function — Method

_build_nn_function(eq, params, sinput)

Build a function that can process a matrix. This is used as a starting point for build_nn_function.

Examples

using SymbolicNeuralNetworks: _build_nn_function, SymbolicNeuralNetwork
using AbstractNeuralNetworks: params, Chain, Dense, NeuralNetwork
import Random
Random.seed!(123)

c = Chain(Dense(2, 1, tanh))
nn = NeuralNetwork(c)
snn = SymbolicNeuralNetwork(nn)
eq = c(snn.input, params(snn))
built_function = _build_nn_function(eq, params(snn), snn.input)
built_function([1. 2.; 3. 4.], params(nn), 1)

# output

1-element Vector{Float64}:
 0.9912108161055604

Note that we have to supply an extra argument (index) to _build_nn_function that we do not have to supply to build_nn_function.

Keyword Arguments

cse: perform common subexpression elimination (default true).

Symbolics stores an expression as a hash-consed directed acyclic graph, but Symbolics.build_function prints it as a tree. Every time a subexpression is reused — the output of layer $n$ feeding each neuron of layer $n+1$, or the forward pass shared by every block of a symbolic pullback — the whole subtree is emitted again, so both the size of the generated code and the amount of redundant arithmetic grow exponentially with the depth of the network. With cse = true the graph is emitted as a let block of intermediate bindings instead, which keeps code size proportional to the number of distinct nodes.

Pass cse = false to recover the old (fully inlined) output; this is mostly useful for debugging, and for very small networks where the binding overhead is not amortized.

Implementation

This first calls Symbolics.build_function with the keyword argument expression = Val{true} and then modifies the generated code by calling:

fix_create_array,

rewrite_arguments,

modify_input_arguments,

fix_map_reduce.

See the docstrings for those functions for details on how the code is modified.

SymbolicNeuralNetworks._build_nn_function_iip — Method

_build_nn_function_iip(eq, params, sinput, soutput; reduce, cse)

The two-input counterpart of _build_nn_function_iip(::EqT, ::NeuralNetworkParameters, ::Symbolics.Arr). The resulting kernel is called with kernel!(out, input, output, ps, k).

Implementation

As for the single-input case, rewrite_arguments2 can be reused unchanged: the ˍ₋out argument that Symbolics.build_function prepends to the in-place signature is not counted in the ˍ₋argN numbering.

SymbolicNeuralNetworks._build_nn_function_iip — Method

_build_nn_function_iip(eq, params, sinput; reduce, cse)

Build an in-place kernel that writes the result for batch column k into a preallocated array:

kernel!(out, input, ps, k)

Returns nothing for scalar-valued equations, for which Symbolics.build_function emits no in-place form.

Implementation

This works like _build_nn_function, but keeps the second (in-place) half of what Symbolics.build_function returns and post-processes it with:

fix_create_array,

rewrite_arguments — unchanged, because the ˍ₋out argument is not counted in the ˍ₋argN numbering,

modify_input_arguments_iip,

fix_map_reduce,

make_kernel_iip.

The in-place code addresses its output with a linear index (ˍ₋out[i] = …) whatever the shape of the equation, which is what lets make_kernel_iip offset the writes by the batch index instead of handing the kernel a view.

SymbolicNeuralNetworks._build_nn_function_per_leaf — Method

_build_nn_function_per_leaf(eqs, sparams, sinput...)

Build one function per entry of eqs. This is the fallback for equation sets that the joint code path of build_nn_function(::Union{NamedTuple, NeuralNetworkParameters}, ::NeuralNetworkParameters, ::Symbolics.Arr...) cannot handle.

Internally this is using function_valued_parameters and apply_element_wise.

SymbolicNeuralNetworks._float_if_integer — Method

_float_if_integer(T)

float(T) for integer types, T for everything else. See allocate_batch_output. Bool is included, as float(Bool) is Float64.

SymbolicNeuralNetworks._get_contents — Method

_get_contents(nt::AbstractArray{<:NamedTuple})

Return the contents of a one-dimensional vector.

Examples

using SymbolicNeuralNetworks: _get_contents

_get_contents([(a = "element_contained_in_vector", )])

# output

(a = "element_contained_in_vector",)

SymbolicNeuralNetworks._get_params — Method

_get_params(ps::NeuralNetworkParameters)

Return the NamedTuple that's equivalent to the NeuralNetworkParameters.

SymbolicNeuralNetworks._iip_batch_wrapper — Method

_iip_batch_wrapper(kernel!, eq_size, reduce)

Allocate the output once and let the in-place kernel! write every batch column into it. eq_size is the size of the (scalarized) equation, which fixes the shape of the result; see allocate_batch_output.

SymbolicNeuralNetworks._iip_batch_wrapper2 — Method

_iip_batch_wrapper2(kernel!, eq_size, reduce)

The two-input counterpart of _iip_batch_wrapper.

SymbolicNeuralNetworks._modify_integer — Method

_modify_integer

If the input is a single integer, subtract 1 from it.

Examples

using SymbolicNeuralNetworks: _modify_integer

s = ["2", "hello", "hello2", "3"]
_modify_integer.(s)

# output
4-element Vector{String}:
 "1"
 "hello"
 "hello2"
 "2"

SymbolicNeuralNetworks._modify_integer2 — Method

_modify_integer2

If the input is a single integer, subtract 2 from it.

Examples

using SymbolicNeuralNetworks: _modify_integer2

s = ["3", "hello", "hello2", "4"]
_modify_integer2.(s)

# output
4-element Vector{String}:
 "1"
 "hello"
 "hello2"
 "2"

SymbolicNeuralNetworks._oop_batch_wrapper — Method

_oop_batch_wrapper(gen_fun, reduce)

Evaluate the out-of-place kernel gen_fun once per batch column and combine the results. Used for scalar-valued equations, for which there is no in-place kernel.

SymbolicNeuralNetworks._oop_batch_wrapper2 — Method

_oop_batch_wrapper2(gen_fun, reduce)

The two-input counterpart of _oop_batch_wrapper.

SymbolicNeuralNetworks._reduce — Method

_reduce(a)

Pick the out-of-place half of what Symbolics.build_function returns. It returns a (out_of_place, in_place) tuple for array-valued equations and a single expression for scalar-valued ones. See _reduce_iip.

SymbolicNeuralNetworks._reduce_iip — Method

_reduce_iip(a)

Pick the in-place half of what Symbolics.build_function returns, or nothing if there is none (which is the case for scalar-valued equations). See _reduce.

SymbolicNeuralNetworks.allocate_batch_output — Method

allocate_batch_output(T, eq_size, batch_size, reduce)

Allocate the result of evaluating an equation of size eq_size over batch_size columns.

The shape matches what Base.reduce(reduce, ::Vector) over the per-column results used to produce:

reduce = +: the per-column results are summed, so the result has the size of the equation.

reduce = hcat, vector-valued equation of length $m$: an $m\times{}N$ matrix.

reduce = hcat, matrix-valued equation of size $(m, n)$: the blocks are placed next to each other, giving an $m\times(n\cdot{}N)$ matrix.

Equations of rank three or higher are not covered by that correspondence — hcat concatenates those along their second dimension and keeps the third, whereas the linear indexing of the in-place kernel flattens everything past the first dimension. No such equation arises in this package.

T is widened with float when it is an integer type: it comes from promoted_eltype, i.e. from the inputs, and an equation over integer inputs generally does not evaluate to an integer.

SymbolicNeuralNetworks.allocate_single_output — Method

allocate_single_output(T, eq_size, reduce)

Allocate the result of evaluating an equation of size eq_size for a single (vector) input. Unlike allocate_batch_output this keeps the shape of the equation, as the output may itself be a matrix.

SymbolicNeuralNetworks.apply_element_wise — Method

apply_element_wise(ps, params, input...)

Apply a function element-wise. ps is an Array where each entry of the array is are NeuralNetworkParameters that store functions. See apply_element_wise(::NeuralNetworkParameters, ::NeuralNetworkParameters, ::Any).

Examples

Vector:

using SymbolicNeuralNetworks: apply_element_wise
using AbstractNeuralNetworks: NeuralNetworkParameters

# parameter values
params = NeuralNetworkParameters((a = 1., b = 2.))
ps = [NeuralNetworkParameters((val1 = (input, params) -> input .+ params.a, val2 = (input, params) -> input .+ params.b))]
apply_element_wise(ps, params, [1.])

# output

1-element Vector{NeuralNetworkParameters{(:val1, :val2), Tuple{Vector{Float64}, Vector{Float64}}}}:
 NeuralNetworkParameters{(:val1, :val2), Tuple{Vector{Float64}, Vector{Float64}}}((val1 = [2.0], val2 = [3.0]))

Matrix:

using SymbolicNeuralNetworks: apply_element_wise
using AbstractNeuralNetworks: NeuralNetworkParameters

# parameter values
params = NeuralNetworkParameters((a = 1., b = 2.))
sc_ps = NeuralNetworkParameters((val1 = (input, params) -> input .+ params.a, val2 = (input, params) -> input .+ params.b))
ps = [sc_ps sc_ps]
apply_element_wise(ps, params, [1.]) |> typeof

# output

Matrix{NeuralNetworkParameters{(:val1, :val2), Tuple{Vector{Float64}, Vector{Float64}}}} (alias for Array{NeuralNetworkParameters{(:val1, :val2), Tuple{Array{Float64, 1}, Array{Float64, 1}}}, 2})

Implementation

This is generating a @generated function.

SymbolicNeuralNetworks.apply_element_wise — Method

apply_element_wise(ps, params, input...)

Apply a function element-wise. ps is a NeuralNetworkParameters-valued function.

Examples

using SymbolicNeuralNetworks: apply_element_wise
using AbstractNeuralNetworks: NeuralNetworkParameters

# parameter values
params = NeuralNetworkParameters((a = 1., b = 2.))
ps = NeuralNetworkParameters((val1 = (input, params) -> input + params.a, val2 = (input, params) -> input + params.b))
apply_element_wise(ps, params, 1.)

# output

NeuralNetworkParameters{(:val1, :val2), Tuple{Float64, Float64}}((val1 = 2.0, val2 = 3.0))

Implementation

This is generating a @generated function.

SymbolicNeuralNetworks.build_function_generated — Method

build_function_generated(reducer, sc_eq, args...; cse)

Call Symbolics.build_function and pick the relevant half of its output with reducer (_reduce for the out-of-place form, _reduce_iip for the in-place one).

Unsupported: reductions over un-scalarized symbolic arrays

Expressions containing an arrayop — a reduction over a Symbolics.Arr that has not been scalarized, e.g. sum(c(input, ps)) — generate code that refers to variables nothing binds, so the resulting function throws an UndefVarError when it is called. This is independent of cse (the two modes just mangle different names). Reduce over collect(c(input, ps)) instead.

SymbolicNeuralNetworks.build_nn_function — Method

build_nn_function(eqs::AbstractArray{<:NeuralNetworkParameters}, sparams, sinput...)

Build an executable function based on an array of symbolic equations eqs.

Examples

using SymbolicNeuralNetworks: build_nn_function, SymbolicNeuralNetwork
using AbstractNeuralNetworks: Chain, Dense, NeuralNetwork, params
import Random
Random.seed!(123)

ch = Chain(Dense(2, 1, tanh))
nn = NeuralNetwork(ch)
snn = SymbolicNeuralNetwork(nn)
eqs = [(a = ch(snn.input, params(snn)), b = ch(snn.input, params(snn)).^2), (c = ch(snn.input, params(snn)).^3, )]
funcs = build_nn_function(eqs, params(snn), snn.input)
input = [1., 2.]
funcs_evaluated = funcs(input, params(nn))

# output

2-element Vector{NamedTuple}:
 (a = [0.985678060655224], b = [0.9715612392570434])
 (c = [0.9576465981186686],)

SymbolicNeuralNetworks.build_nn_function — Method

build_nn_function(eqs, nn, soutput)

Build an executable function that can also depend on an output. The resulting built_function is then called with:

built_function(input, output, ps)

Also compare this to build_nn_function(::EqT, ::AbstractSymbolicNeuralNetwork).

Keyword Arguments

cse: perform common subexpression elimination when generating code (default true).

inplace: evaluate a batch with an in-place kernel (default true). As for the single-input case the default result mutates a preallocated array and can therefore not be differentiated with Zygote; see build_nn_function(::EqT, ::AbstractSymbolicNeuralNetwork).

Extended Help

See the extended help section of build_nn_function(::EqT, ::AbstractSymbolicNeuralNetwork).

SymbolicNeuralNetworks.build_nn_function — Method

build_nn_function(eqs::Union{NamedTuple, NeuralNetworkParameters}, sparams, sinput...)

Return a function that takes an input, (optionally) an output and neural network parameters and returns a NeuralNetworkParameters-valued output.

Examples

using SymbolicNeuralNetworks: build_nn_function, SymbolicNeuralNetwork
using AbstractNeuralNetworks: Chain, Dense, NeuralNetwork, params
import Random
Random.seed!(123)

c = Chain(Dense(2, 1, tanh))
nn = NeuralNetwork(c)
snn = SymbolicNeuralNetwork(nn)
eqs = (a = c(snn.input, params(snn)), b = c(snn.input, params(snn)).^2)
funcs = build_nn_function(eqs, params(snn), snn.input)
input = [1., 2.]
funcs_evaluated = funcs(input, params(nn))

# output

(a = [0.985678060655224], b = [0.9715612392570434])

Implementation

All the entries of eqs are generated as a single function, whose flat output is split up again afterwards; see flatten_eqs and unflatten. Generating one function per entry instead would re-derive everything the entries have in common — for a symbolic pullback that is the whole forward pass, once per parameter array — and would compile one RuntimeGeneratedFunction per entry rather than one in total.

Equation sets that contain a scalar-valued entry fall back to one function per entry (via function_valued_parameters and apply_element_wise), because Symbolics.build_function emits no in-place form for scalars.

Joint code generation is independent of the inplace keyword: inplace = false still generates the whole set as one function, it just evaluates it out of place (and is then Zygote-differentiable; see build_nn_function(::EqT, ::AbstractSymbolicNeuralNetwork)).

SymbolicNeuralNetworks.build_nn_function — Method

build_nn_function(eq, nn)

Build an executable function based on a symbolic equation, a symbolic input array and a SymbolicNeuralNetwork.

This function can be called with:

built_function(input, ps)

Keyword Arguments

cse: perform common subexpression elimination when generating code (default true). See _build_nn_function.

inplace: evaluate a batch with an in-place kernel (default true). See below.

The default result cannot be differentiated with `Zygote`

With inplace = true the returned function allocates its result and lets the generated kernel mutate it, which Zygote does not support (Mutating arrays is not supported). Pass inplace = false to get the out-of-place version, which evaluates the kernel once per batch column and combines the results with Base.reduce; that one is differentiable, but allocates an array per column. Forward-mode AD (ForwardDiff) works with either, as the element type of the preallocated array is promoted over the inputs (see promoted_eltype).

Implementation

Internally this is calling _build_nn_function_iip and then parallelizing the expression via the index k. The kernel writes straight into a preallocated output array, so evaluating a batch costs a single allocation instead of one array per column plus a Base.reduce fold.

For scalar-valued equations Symbolics.build_function does not emit an in-place form; those fall back to the out-of-place _build_nn_function, as does inplace = false.

Extended Help

The functions mentioned in the implementation section were adjusted ad-hoc to deal with problems that emerged on the fly. Other problems may occur. In case you bump into one please open an issue on github.

SymbolicNeuralNetworks.derivative — Method

derivative(g)

Examples

We compare this to symbolic_pullback here:

using SymbolicNeuralNetworks: SymbolicNeuralNetwork, Gradient, derivative, symbolic_pullback
using AbstractNeuralNetworks

c = Chain(Dense(2, 1, tanh))
nn = SymbolicNeuralNetwork(c)
g = Gradient(nn)
∇ = derivative(g)

isequal(∇, symbolic_pullback(g.f, nn))

# output

true

SymbolicNeuralNetworks.fix_create_array — Method

fixcreatearray(s)

Fix a problem that occurs in connection with create_array.

The function create_array from SymbolicUtils.Code takes as first input the type of a symbolic array. For reasons that are not entirely clear yet the first argument of create_array ends up being ˍ₋arg2, which is a NamedTuple of symoblic arrays. We solve this problem by replacing typeof(ˍ₋arg[0-9]+) with Array, which is the most generic input to create_array and avoids MethodErrors when sinput is a non-standard array type such as ReshapedArray or SubArray.

Examples

using SymbolicNeuralNetworks: fix_create_array

s = "(SymbolicUtils.Code.create_array)(typeof(ˍ₋arg2)"
fix_create_array(s)

# output

"SymbolicUtils.Code.create_array(Array"

Implementation

This is used for _build_nn_function(::EqT, ::NeuralNetworkParameters, ::Symbolics.Arr) and _build_nn_function(::EqT, ::NeuralNetworkParameters, ::Symbolics.Arr, ::Symbolics.Arr).

SymbolicNeuralNetworks.fix_map_reduce — Method

fix_map_reduce(s)

Replace Symbolics._mapreduce with mapreduce (from Base).

When we generate a function with Symbolics.build_function it often contains Symbolics._mapreduce which cannot be differentiated with Zygote. We get around this by replacing Symbolics._mapreduce with mapreduce and also doing:

replace(s, ", Colon(), (:init => false,)" => ", dims = Colon()")

Implementation

This is used for _build_nn_function(::EqT, ::NeuralNetworkParameters, ::Symbolics.Arr) and _build_nn_function(::EqT, ::NeuralNetworkParameters, ::Symbolics.Arr, ::Symbolics.Arr).

SymbolicNeuralNetworks.flatten_eqs — Method

flatten_eqs(eqs)

Concatenate every entry of eqs into one vector of scalar equations, together with a template: a copy of the nested structure of eqs in which each entry has been replaced by the FlatSlice describing where it went. unflatten reverses this.

Returns nothing if any entry is scalar-valued, which the joint code path cannot handle.

Examples

using SymbolicNeuralNetworks: flatten_eqs, SymbolicNeuralNetwork
using AbstractNeuralNetworks: Chain, Dense, params

c = Chain(Dense(2, 3, tanh))
snn = SymbolicNeuralNetwork(c)
flat, template = flatten_eqs((a = c(snn.input, params(snn)), b = c(snn.input, params(snn)).^2))
(length(flat), template.a.range, template.b.range)

# output

(6, 1:3, 4:6)

SymbolicNeuralNetworks.function_valued_parameters — Method

function_valued_parameters(eqs::Union{NamedTuple, NeuralNetworkParameters}, sparams, sinput...)

Return an executable function for each entry in eqs. This still has to be processed with apply_element_wise.

Examples

using SymbolicNeuralNetworks: function_valued_parameters, SymbolicNeuralNetwork
using AbstractNeuralNetworks: Chain, Dense, NeuralNetwork, params
import Random
Random.seed!(123)

c = Chain(Dense(2, 1, tanh))
nn = NeuralNetwork(c)
snn = SymbolicNeuralNetwork(nn)
eqs = (a = c(snn.input, params(snn)), b = c(snn.input, params(snn)).^2)
funcs = function_valued_parameters(eqs, params(snn), snn.input)
input = [1., 2.]
ps = params(nn)
a = c(input, ps)
b = c(input, ps).^2

(funcs.a(input, ps), funcs.b(input, ps)) .≈ (a, b)

# output

(true, true)

SymbolicNeuralNetworks.make_kernel — Method

Examples

using SymbolicNeuralNetworks

s = "function (sinput, ps)\n begin\n getindex(sinput, 1) + getindex(sinput, 2) \n end\n end"
SymbolicNeuralNetworks.make_kernel(s)

# output

"function (sinput, ps, k)\n begin\n getindex(sinput, 1, k) + getindex(sinput, 2, k) \n end\n end"

SymbolicNeuralNetworks.make_kernel2 — Method

Examples

using SymbolicNeuralNetworks

s = "function (sinput, soutput, ps)\n begin\n getindex(sinput, 1) + getindex(soutput, 2) \n end\n end"
SymbolicNeuralNetworks.make_kernel2(s)

# output

"function (sinput, soutput, ps, k)\n begin\n getindex(sinput, 1, k) + getindex(soutput, 2, k) \n end\n end"

SymbolicNeuralNetworks.make_kernel_iip — Method

make_kernel_iip(s, reduce, eq_length)

The in-place counterpart of make_kernel: add the batch index k to the arguments, index sinput with it, and redirect the writes to ˍ₋out so that a single preallocated array can hold the result for the whole batch.

eq_length is the number of entries of the (scalarized) equation. The generated code assigns each of them exactly once, which redirect_output_writes checks.

Which redirection is applied depends on how the per-column results are combined:

reduce = +: the writes become += and every column accumulates into the same buffer (which allocate_batch_output zeroes).

reduce = hcat: the writes are shifted by $(k - 1)\cdot\mathrm{eq\_length}$, which is exactly the offset of block k in the column-major layout of the concatenated result.

Examples

using SymbolicNeuralNetworks: make_kernel_iip

s = "function (ˍ₋out, sinput, ps)\n begin\n ˍ₋out[1] = getindex(sinput, 2)\n ˍ₋out[2] = getindex(sinput, 1) \n end\n end"
make_kernel_iip(s, hcat, 2)

# output

"function (ˍ₋out, sinput, ps, k)\n begin\n ˍ₋out[1 + (k - 1) * 2] = getindex(sinput, 2, k)\n ˍ₋out[2 + (k - 1) * 2] = getindex(sinput, 1, k) \n end\n end"

using SymbolicNeuralNetworks: make_kernel_iip

s = "function (ˍ₋out, sinput, ps)\n begin\n ˍ₋out[1] = getindex(sinput, 2)\n ˍ₋out[2] = getindex(sinput, 1) \n end\n end"
make_kernel_iip(s, +, 2)

# output

"function (ˍ₋out, sinput, ps, k)\n begin\n ˍ₋out[1] += getindex(sinput, 2, k)\n ˍ₋out[2] += getindex(sinput, 1, k) \n end\n end"

SymbolicNeuralNetworks.make_kernel_iip2 — Method

make_kernel_iip2(s, reduce, eq_length)

The two-input counterpart of make_kernel_iip.

Examples

using SymbolicNeuralNetworks: make_kernel_iip2

s = "function (ˍ₋out, sinput, soutput, ps)\n begin\n ˍ₋out[1] = getindex(sinput, 1) + getindex(soutput, 2)\n ˍ₋out[2] = getindex(sinput, 2) \n end\n end"
make_kernel_iip2(s, +, 2)

# output

"function (ˍ₋out, sinput, soutput, ps, k)\n begin\n ˍ₋out[1] += getindex(sinput, 1, k) + getindex(soutput, 2, k)\n ˍ₋out[2] += getindex(sinput, 2, k) \n end\n end"

SymbolicNeuralNetworks.modify_input_arguments — Method

modify_input_arguments(s)

Change input arguments of type (sinput, ps.L1, ps.L2) etc to (sinput, ps). This should be used after rewrite_arguments. Also see build_nn_function.

Examples

using SymbolicNeuralNetworks: modify_input_arguments

s = "(sinput, ps.L1, ps.L2, ps.L3)"
modify_input_arguments(s)

# output
"(sinput, ps)"

SymbolicNeuralNetworks.modify_input_arguments2 — Method

modify_input_arguments2(s)

Change input arguments of type (sinput, soutput, ps.L1, ps.L2) etc to (sinput, soutput, ps). This should be used after rewrite_arguments.

Examples

using SymbolicNeuralNetworks: modify_input_arguments2

s = "(sinput, soutput, ps.L1, ps.L2, ps.L3)"
modify_input_arguments2(s)

# output
"(sinput, soutput, ps)"

SymbolicNeuralNetworks.modify_input_arguments_iip — Method

modify_input_arguments_iip(s)

Change input arguments of type (ˍ₋out, sinput, ps.L1, ps.L2) etc to (ˍ₋out, sinput, ps).

This is the in-place counterpart of modify_input_arguments and should be used after rewrite_arguments. See _build_nn_function_iip.

Examples

using SymbolicNeuralNetworks: modify_input_arguments_iip

s = "(ˍ₋out, sinput, ps.L1, ps.L2, ps.L3)"
modify_input_arguments_iip(s)

# output
"(ˍ₋out, sinput, ps)"

SymbolicNeuralNetworks.modify_input_arguments_iip2 — Method

modify_input_arguments_iip2(s)

Change input arguments of type (ˍ₋out, sinput, soutput, ps.L1, ps.L2) etc to (ˍ₋out, sinput, soutput, ps). See modify_input_arguments_iip.

Examples

using SymbolicNeuralNetworks: modify_input_arguments_iip2

s = "(ˍ₋out, sinput, soutput, ps.L1, ps.L2, ps.L3)"
modify_input_arguments_iip2(s)

# output
"(ˍ₋out, sinput, soutput, ps)"

SymbolicNeuralNetworks.promoted_eltype — Method

promoted_eltype(args...)

The element type the generated code will produce, promoted over all inputs.

This is needed because the in-place kernels write into an array that we have to allocate before calling them, so the element type cannot be inferred from a result. Promoting over the inputs keeps Float32 parameters, symbolic (Num) inputs and ForwardDiff.Dual numbers working.

Note that this derives the element type from the inputs, not from the expression, which the out-of-place path (_build_nn_function) does instead. The two can differ: an equation over integer inputs and integer parameters evaluates to a Float64, which no Array{Int} can hold. The allocators therefore widen an integer type with float (see allocate_batch_output). A Float32 network whose generated code contains a Float64 constant is rounded to Float32 rather than widened, which is the behaviour one wants for the network but is worth being aware of.

SymbolicNeuralNetworks.redirect_output_writes — Method

redirect_output_writes(s, reduce, eq_length)

Rewrite the ˍ₋out[i] = … assignments of in-place generated code. See make_kernel_iip.

Implementation

Unlike the other rewrites in the pipeline this one cannot fail loudly on its own: a replace that matches nothing returns the string unchanged, and the kernel then still compiles and runs. It would just write every batch column into the same place (reduce = hcat) or overwrite instead of accumulate (reduce = +), i.e. silently return wrong numbers. We therefore check that there is exactly one assignment per entry of the equation, which is the property test/build_function/codegen_drift.jl guards upstream.

SymbolicNeuralNetworks.rewrite_arguments — Method

rewrite_arguments(s)

Replace ˍ₋arg2, ˍ₋arg3, ... with ps.L1, ps.L2 etc. This is used after Symbolics.build_function.

Examples

using SymbolicNeuralNetworks: rewrite_arguments
s = "We test if strings that contain ˍ₋arg2 and ˍ₋arg3 can be converted in the right way."
rewrite_arguments(s)

# output
"We test if strings that contain ps.L1 and ps.L2 can be converted in the right way."

Implementation

The input is first split at the relevant points and then we call _modify_integer. The routine _modify_integer ensures that we start counting at 1 and not at 2. By defaut the arguments of the generated function that we get after applying Symbolics.build_function are (x, ˍ₋arg2, ˍ₋arg3) etc. We first change this to (x, ps.L2, ps.L3) etc. and then to (x, ps.L1, ps.L2) etc. via _modify_integer.

SymbolicNeuralNetworks.rewrite_arguments2 — Method

rewrite_arguments2(s)

Replace ˍ₋arg3, ˍ₋arg4, ... with ps.L1, ps.L2 etc. Note that we subtract two from the input, unlike rewrite_arguments where it is one.

Examples

using SymbolicNeuralNetworks: rewrite_arguments2
s = "We test if strings that contain ˍ₋arg3 and ˍ₋arg4 can be converted in the right way."
rewrite_arguments2(s)

# output
"We test if strings that contain ps.L1 and ps.L2 can be converted in the right way."

Implementation

The input is first split at the relevant points and then we call _modify_integer2. The routine _modify_integer2 ensures that we start counting at 1 and not at 3. See rewrite_arguments.

SymbolicNeuralNetworks.symbolic_pullback — Method

symbolic_pullback(f, nn)

This takes a symbolic fthat depends on the parameters innn` and returns the corresponding pullback (a symbolic expression).

This is used by Gradient and SymbolicPullback.

Examples

using SymbolicNeuralNetworks: SymbolicNeuralNetwork, symbolic_pullback
using AbstractNeuralNetworks
using AbstractNeuralNetworks: params
using LinearAlgebra: norm

c = Chain(Dense(2, 1, tanh))
nn = SymbolicNeuralNetwork(c)
output = c(nn.input, params(nn))
spb = symbolic_pullback(output, nn)

spb[1].L1.b

# output

1-element Vector{Symbolics.Num}:
 1 - (tanh(W_2[1] + W_1[1, 1]*sinput[1] + W_1[1, 2]*sinput[2])^2)

SymbolicNeuralNetworks.symboliccounter! — Method

symboliccounter!(cache, arg; redundancy)

Add a specific argument to the cache.

Examples

using SymbolicNeuralNetworks: symboliccounter!

cache = Dict()
var = symboliccounter!(cache, :var)
(cache, var)

# output
(Dict{Any, Any}(:var => 1), :var_1)

SymbolicNeuralNetworks.symbolize! — Function

symbolize!(cache, nt, var_name)

Symbolize all the arguments in nt.

Examples

using SymbolicNeuralNetworks: symbolize!

cache = Dict()
sym = symbolize!(cache, .1, :X)
(sym, cache)

# output

(X_1, Dict{Any, Any}(:X => 1))

using SymbolicNeuralNetworks: symbolize!

cache = Dict()
arr = rand(2, 1)
sym_scalar = symbolize!(cache, .1, :X)
sym_array = symbolize!(cache, arr, :Y)
(sym_array, cache)

# output

(Y_1[1:2,1:1], Dict{Any, Any}(:X => 1, :Y => 1))

Note that the for the second case the cache is storing a scalar under :X and an array under :Y. If we use the same label for both we get:

using SymbolicNeuralNetworks: symbolize!

cache = Dict()
arr = rand(2, 1)
sym_scalar = symbolize!(cache, .1, :X)
sym_array = symbolize!(cache, arr, :X)
(sym_array, cache)

# output

(X_2[1:2,1:1], Dict{Any, Any}(:X => 2))

We can also use symbolize! with NamedTuples:

using SymbolicNeuralNetworks: symbolize!

cache = Dict()
nt = (a = 1, b = [1, 2])
sym = symbolize!(cache, nt, :X)
(sym, cache)

# output

((a = X_1, b = X_2[1:2]), Dict{Any, Any}(:X => 2))

And for neural network parameters:

using SymbolicNeuralNetworks: symbolize!
using AbstractNeuralNetworks: NeuralNetwork, params, Chain, Dense

nn = NeuralNetwork(Chain(Dense(1, 2; use_bias = false), Dense(2, 1; use_bias = false)))
cache = Dict()
sym = symbolize!(cache, params(nn), :X) |> typeof

# output

AbstractNeuralNetworks.NeuralNetworkParameters{(:L1, :L2), Tuple{@NamedTuple{W::Symbolics.Arr{Symbolics.Num, 2}}, @NamedTuple{W::Symbolics.Arr{Symbolics.Num, 2}}}}

Implementation

Internally this is using symboliccounter!. This function is also adjusting/altering the cache (that is optionally supplied as an input argument).

SymbolicNeuralNetworks.unflatten — Method

unflatten(template, out)

Split the flat result out of a jointly generated function back into the nested structure recorded in template. The inverse of flatten_eqs.

Implementation

Each entry is copied out of out rather than viewed into it, so that the entries are ordinary Arrays (as they were when every entry had its own generated function) and cannot alias each other.

out is indexed by its number of dimensions, which is how the layout of the batch is accounted for:

a vector when the per-column results were summed (reduce = +), in which case every entry keeps the shape of its equation,

a $P\times{}N$ matrix when they were concatenated (reduce = hcat), in which case an entry of size $(m, n, \ldots)$ becomes an $m\times(n\cdot\ldots\cdot{}N)$ matrix, exactly as Base.reduce(hcat, …) would have produced.